+++ sample/fixed-elem/sample.js
@@ -1,0 +1,6 @@
+jQuery(document).ready(function($) {
+  $('aside').simpleScrollFollow({
+    upper_side: '#menu-fixed-top',
+    lower_side: '#menu-fixed-bottom'
+  });
+});