+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # CHANGELOG jquery.simple-scroll-follow
+
+## v3.1.0
+#### Add option `upper_side` and `lower_side`
+If you use fixed menu such as [Bootstrap .navbar-fixed-top](http://getbootstrap.com/components/#navbar-fixed-top), the options below will help you.
+
+```javascript
+$('#foo').simpleScrollFollow({
+  upper_side: '#menu-fixed-top',
+  lower_side: '#menu-fixed-bottom'
+});
+```
+
+- - -
 ## v3.0.0
 #### Option "instance" is deleted.
 In v3.x, simpleScrollFollow() always returns jQuery object, so jQuery method chaining always works.